Move ImageGradient paint logic to the runtime and expose via API.
