+++ Gems/GradientSignal/Code/Source/Editor/EditorImageGradientComponent.cpp
@@ -661,17 +661,25 @@ namespace GradientSignal
         return true;
     }
 
-    void EditorImageGradientComponent::StartImageModification()
+    AZ::EntityComponentIdPair EditorImageGradientComponent::StartImageModification()
     {
         // While we're editing, we need to set all the configuration properties to read-only and refresh them.
         // Otherwise, the property changes could conflict with the current painted modifications.
         m_configuration.m_imageModificationActive = true;
         AzToolsFramework::ToolsApplicationEvents::Bus::Broadcast(
             &AzToolsFramework::ToolsApplicationEvents::InvalidatePropertyDisplay, AzToolsFramework::Refresh_AttributesAndValues);
+
+        ImageGradientModificationBus::Event(GetEntityId(), &ImageGradientModifications::StartImageModification);
+
+        // Return the runtime entity/component ID pair. This is used to hook the component mode's paintbrush to the runtime component
+        // instead of the editor component so that it can modify the image successfully.
+        return { m_component.GetEntityId(), m_component.GetId() };
     }
 
     void EditorImageGradientComponent::EndImageModification()
     {
+        ImageGradientModificationBus::Event(GetEntityId(), &ImageGradientModifications::EndImageModification);
+
         // We're done editing, so set all the configuration properties back to writeable and refresh them.
         m_configuration.m_imageModificationActive = false;
         AzToolsFramework::ToolsApplicationEvents::Bus::Broadcast(
