+++ Gems/GradientSignal/Code/Source/Editor/EditorImageGradientComponentMode.h
@@ -10,10 +10,10 @@
 
 #include <AzToolsFramework/ComponentMode/EditorBaseComponentMode.h>
 #include <AzToolsFramework/Manipulators/PaintBrushManipulator.h>
-#include <AzFramework/PaintBrush/PaintBrushNotificationBus.h>
 #include <AzToolsFramework/PaintBrush/PaintBrushSubModeCluster.h>
 #include <AzToolsFramework/Undo/UndoSystem.h>
 #include <AzToolsFramework/ViewportUi/ViewportUiRequestBus.h>
+#include <GradientSignal/Ebuses/ImageGradientModificationBus.h>
 
 namespace GradientSignal
 {
@@ -22,66 +22,27 @@ namespace GradientSignal
 
     class EditorImageGradientComponentMode
         : public AzToolsFramework::ComponentModeFramework::EditorBaseComponentMode
-        , private AzFramework::PaintBrushNotificationBus::Handler
+        , private ImageGradientModificationNotificationBus::Handler
     {
     public:
         EditorImageGradientComponentMode(const AZ::EntityComponentIdPair& entityComponentIdPair, AZ::Uuid componentType);
         ~EditorImageGradientComponentMode() override;
 
-        // EditorBaseComponentMode
+        // EditorBaseComponentMode overrides...
         void Refresh() override;
         AZStd::vector<AzToolsFramework::ActionOverride> PopulateActionsImpl() override;
         bool HandleMouseInteraction(const AzToolsFramework::ViewportInteraction::MouseInteractionEvent& mouseInteraction) override;
         AZStd::string GetComponentModeName() const override;
 
     protected:
-        // PaintBrushNotificationBus overrides
-        void OnBrushStrokeBegin(const AZ::Color& color) override;
-        void OnBrushStrokeEnd() override;
-        void OnPaint(const AZ::Aabb& dirtyArea, ValueLookupFn& valueLookupFn, BlendFn& blendFn) override;
-        void OnSmooth(
-            const AZ::Aabb& dirtyArea,
-            ValueLookupFn& valueLookupFn,
-            AZStd::span<const AZ::Vector3> valuePointOffsets,
-            SmoothFn& smoothFn) override;
-        AZ::Color OnGetColor(const AZ::Vector3& brushCenter) override;
-
-        void OnPaintSmoothInternal(
-            const AZ::Aabb& dirtyArea, ValueLookupFn& valueLookupFn,
-            AZStd::function<float(const AZ::Vector3& worldPosition, float gradientValue, float opacity)> combineFn);
+        // ImageGradientModificationNotificationBus overrides...
+        void OnImageGradientBrushStrokeBegin() override;
+        void OnImageGradientBrushStrokeEnd(AZStd::shared_ptr<ImageTileBuffer> changedDataBuffer, const AZ::Aabb& dirtyRegion) override;
 
         void BeginUndoBatch();
         void EndUndoBatch();
 
     private:
-        struct PaintStrokeData
-        {
-            //! A buffer to accumulate a single paint stroke into. This buffer is used to ensure that within a single paint stroke,
-            //! we only perform an operation on a pixel once, not multiple times.
-            //! After the paint stroke is complete, this buffer is handed off to the undo/redo batch so that we can undo/redo each stroke.
-            AZStd::unique_ptr<ImageTileBuffer> m_strokeBuffer;
-
-            //! The meters per pixel in each direction for this image gradient.
-            //! These help us query the paintbrush for exactly one world position per image pixel.
-            float m_metersPerPixelX = 0.0f;
-            float m_metersPerPixelY = 0.0f;
-
-            //! The intensity of the paint stroke (0 - 1)
-            float m_intensity = 0.0f;
-
-            //! The opacity of the paint stroke (0 - 1)
-            float m_opacity = 0.0f;
-
-            //! Track the dirty region for each paint stroke so that we can store it in the undo/redo buffer
-            //! to send with change notifications.
-            AZ::Aabb m_dirtyRegion = AZ::Aabb::CreateNull();
-        };
-
-        PaintStrokeData m_paintStrokeData;
-
-        //! The entity/component that owns this paintbrush.
-        AZ::EntityComponentIdPair m_ownerEntityComponentId;
-
         //! The core paintbrush manipulator and painting logic.
         AZStd::shared_ptr<AzToolsFramework::PaintBrushManipulator> m_brushManipulator;
