+++ Gems/GradientSignal/Code/gradientsignal_files.cmake
@@ -17,6 +17,7 @@ set(FILES
     Include/GradientSignal/Components/GradientSurfaceDataComponent.h
     Include/GradientSignal/Components/GradientTransformComponent.h
     Include/GradientSignal/Components/ImageGradientComponent.h
+    Include/GradientSignal/Components/ImageGradientModification.h
     Include/GradientSignal/Components/InvertGradientComponent.h
     Include/GradientSignal/Components/LevelsGradientComponent.h
     Include/GradientSignal/Components/MixedGradientComponent.h
@@ -60,6 +61,7 @@ set(FILES
     Source/Components/GradientSurfaceDataComponent.cpp
     Source/Components/GradientTransformComponent.cpp
     Source/Components/ImageGradientComponent.cpp
+    Source/Components/ImageGradientModification.cpp
     Source/Components/InvertGradientComponent.cpp
     Source/Components/LevelsGradientComponent.cpp
     Source/Components/MixedGradientComponent.cpp
